+++ SwiftUI_Pull_to_Refresh/Views/City/WeatherList/Forecast/ForecastView.swift
@@ -40,57 +40,21 @@ struct ForecastView: View {
 			
 			// Temperature.
 			HStack {
-				Text(viewModel.temperatureString)
+				Text(viewModel.celsiusString)
 					.largeStyle()
 				Spacer()
 			}
 			.frame(width: 80)
 			
 			// Temperature bar.
-			HStack(spacing: 2) {
-				RoundedRectangle(cornerRadius: 2)
-					.foregroundColor(UI.Color.gray.opacity(0.2))
-					.frame(height: 2)
-					.overlay(
-						GeometryReader { geometry in
-							HStack {
-								Spacer()
-								RoundedRectangle(cornerRadius: 2)
-									.foregroundColor(UI.Color.gray)
-									.frame(
-										width: geometry.size.width * negativePercentage,
-										height: 2
-									)
-							}
-						}
-					)
-				RoundedRectangle(cornerRadius: 2)
-					.foregroundColor(UI.Color.gray.opacity(0.2))
-					.frame(height: 2)
-					.overlay(
-						GeometryReader { geometry in
-							HStack {
-								RoundedRectangle(cornerRadius: 2)
-									.foregroundColor(UI.Color.green)
-									.frame(
-										width: geometry.size.width * positivePercentage,
-										height: 2
-									)
-								Spacer()
-							}
-						}
-					)
-			}
+			TermperatureBarView(
+				temperature: viewModel.temperature,
+				smallestTemperature: viewModel.smallestTemperature,
+				greatestTemperature: viewModel.greatestTemperature
+			)
+			
 		}
 		.frame(height: UI.rowHeight)
 		.padding(.horizontal, UI.padding)
-	}
-	
-	var negativePercentage: CGFloat {
-		0.5
-	}
-	
-	var positivePercentage: CGFloat {
-		0.5
 	}
 }